move convolve to StructuredReader, fully handled there
